refactor(stella-core): split compaction's test suite into compaction/tests.rs
The two new retention-gate witnesses pushed compaction.rs to 1529
lines, past the 1500 hard limit for non-grandfathered files. Move the
whole inline tests module to a child file — the loop_detect.rs
precedent from #1797 — leaving compaction.rs at 724 lines.
